Major reorganization of files and naming
