--- includes/dmmcrm-metaboxes-config.php
@@ -1,12 +1,0 @@
-<?php
-if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
-/**
- * Metaboxes Configuration File
- *
- * This defines the metaboxes. 
- * 
- * @license   GPL-3.0
- * @version   0.0.1
- */
- 
-
